requirements.txt
Django
djangorestframework
markdown
django-filter
coreapi
django-cors-headers